fix docker swarm test
